refactor(chat): extract search state into hook
